Fix for private key not loading + useragent
In my experience I was unable to get this sample IdP code to work.  The
signing certificate's private key was always NULL without adding the
X509KeyStorageFlags.PersistKeySet flag.

Also, a lot of web servers are now configured to return a 403 -
Forbidden response when the UserAgent header is not specified.  I
experienced this exact issue against WordPress running on the hosting
provider "Pressable."  After Googling the issue it seems there is
somewhat of a consensus to return a 403 response when said header is
left off.  Hence, the addition to AddHttpClient() where a default header
of "ITFoxtec.TestIdpCore" is specified for the UserAgent header.