Fix for private key not loading + useragent
In my experience I had a bit of difficulty getting this sample Idp code
to work. The signing certificate's private key was always NULL without
adding the following X509KeyStorageFlags:
  * X509KeyStorageFlags.MachineKeySet
  * X509KeyStorageFlags.PersistKeySet

Also, it seems a lot of web servers are now configured to return a
403 - Forbidden response when the UserAgent header is not
specified. I experienced this exact issue against WordPress running on
the hosting provider "Pressable." Hence, the addition to
AddHttpClient() where a default header of "ITFoxtec.TestIdpCore" is
pecified for the UserAgent header.