Non-generic typing of Saml2Binding
## Background
Prior to this change it was not possible to type a variable as `Saml2Binding` with no knowledge of the implementing type.

So if I want to write a `ResolveBindingForSomething(...)` method then I'm pretty much stuck because of typing limitations.

## Change
The core change is to remove the `<T>` parameter of `Saml2Binding`. To support this I made `BindInternal` agnostic and changed the `Bind` methods into extensions that will organically return the same `T`.

## Result
I can now write code that does not know about underlying types:
```
var binding = ResolveBindingForSomething(...);
binding.ReadSamlResponse(httpRequest, samlResponse);
binding.Unbind(httpRequest, samlResponse);
var relayStateQuery = binding.GetRelayStateQuery();
```